// typings/athlyte/baseball/game/game.d.ts
export interface IPlayerPitchingStats {
  name: string;
  ip: number;
  h: number;  
  er: number;
  bb: number;
  so: number;
  r: number;
  ab: number;
  ground: number;
}


export interface IPlayerHittingStats {
    name: string;
    bats:string;
    bat: string;
    atbat: number;
    runs: number;
    hits:number;
    rbi:number;
    so:number;
    ground:number;
    fly: number;
}